+++ routes/product.js
@@ -28,11 +28,4 @@ router.put(
   require("../controllers/Products/update").process
 );
 
-router.put(
-  "getAllViaCache",
-  userAuthentication,
-  require("../controllers/Caching/getAll").process
-);
-
-router.get("/getAllViaCache", require("../controllers/Caching/getAll").process);
 module.exports = router;
